Rename cities into city
