+++ src/redux/index.js
@@ -2,14 +2,11 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import rootReducer from './reducers';
-import cities from '../assets/samples/cities-sample';
 
 const middleware = [thunk];
 const middlewareEnhancer = applyMiddleware(...middleware);
 
-const initialState = {
-  cities,
-};
+const initialState = {};
 
 const store = createStore(rootReducer, initialState, composeWithDevTools(middlewareEnhancer));
