--- src/redux/reducers/cities.js
@@ -1,18 +1,0 @@
-import { GET_ALL_CITIES } from '../actions/actionsTypes';
-
-const initialState = {
-  cities: [],
-};
-
-const cityReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case GET_ALL_CITIES:
-      return [
-        ...state, action.cities,
-      ];
-    default:
-      return state;
-  }
-};
-
-export default cityReducer;
