+++ src/redux/reducers/index.js
@@ -1,4 +1,4 @@
 import { combineReducers } from 'redux';
-import cityReducer from './cities';
+import cityReducer from './city';
 
 export default combineReducers({ cities: cityReducer });
